project/openclawd_stack/exporter/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pydantic==2.10.6
gspread==6.1.4
google-auth==2.38.0
